First API tested and more small changes

+++ src/models/agency.go
@@ -1,7 +1,7 @@
 package models
 
 type Agency struct {
-	Id 				uint8 	`json:"id"`
+	Id 				int 	`json:"id"`
 	AgencyId 		string 	`json:"agencyId"`
 	AgencyUrl 		string 	`json:"agencyUrl"`
 	AgencyTimezone	string  `json:"agencyTimezone"`
