+++ src/models/fareRule.go
@@ -1,10 +1,10 @@
 package models
 
 type FareRule struct {
-	Id 			uint32 	`json:"id"`
-	FareId 		uint32 	`json:"fareId"`
+	Id 			int 	`json:"id"`
+	FareId 		int 	`json:"fareId"`
 	RouteId 	string 	`json:"routeId"`
-	OriginId 	uint16 	`json:"originId"`
-	DestinationId uint16 `json:"destinationId"`
-	ContainsId 	uint16 	`json:"containsId"`
+	OriginId 	int 	`json:"originId"`
+	DestinationId int `json:"destinationId"`
+	ContainsId 	int 	`json:"containsId"`
 }
