+++ src/models/shape.go
@@ -1,10 +1,10 @@
 package models
 
 type Shape struct{
-	Id 				uint32 		`json:"id"`
+	Id 				int 		`json:"id"`
 	ShapeId 		string 		`json:"shapeId"`
 	ShapePtLat		float32		`json:"shapePtLat"`
 	ShapePtLon		float32 	`json:"shapePtLon"`
-	ShapePtSequence uint16 		`json:"shapePtSequence"`
-	ShapeDistTraveled float32 	`json:"shapeDistTraveled"`
+	ShapePtSequence int 		`json:"shapePtSequence"`
+	ShapeDistTraveled int 		`json:"shapeDistTraveled"`
 }
